+++ lib/src/registry.dart
@@ -112,6 +112,7 @@ final Map<String, String> _defaultFallbackLanguageNames = {
   'objc': BuiltInLanguages.objectiveC.name,
   'obj-c': BuiltInLanguages.objectiveC.name,
   'objective-c': BuiltInLanguages.objectiveC.name,
+  'rb': BuiltInLanguages.ruby.name,
   'clang': BuiltInLanguages.c.name,
   'c++': BuiltInLanguages.cPlusPlus.name,
   'cplusplus': BuiltInLanguages.cPlusPlus.name,